(cleanup) move db.js into server folder
