+++ server/email/emailModel.js
@@ -1,4 +1,4 @@
-require('../../db/db.js');
+require('../db.js');
 var mongoose = require('mongoose');
 
 var EmailSchema = new mongoose.Schema({
