+++ server/marketing/marketingModel.js
@@ -1,4 +1,4 @@
-require('../../db/db.js');
+require('../db.js');
 var mongoose = require('mongoose');
 
 var MarketingSchema = new mongoose.Schema({
